+++ CHANGELOG.md
@@ -55,6 +55,7 @@ _Breaking developer changes, which may affect downstream projects or sites that 
 * Skip `disconnected_way` validation for Golf Paths ([#11863], thanks [@Kayd-06])
 * Also suggest to remove deprecated tag when the object already has the corresponding new tag ([#9272])
 * Add explanation how to fix "point should be vertex" validation warning ([#10958])
+* Suggest to "reposition features" before other fixes when the validator found overlapping buildings ([#11329])
 #### :bug: Bugfixes
 * Fix the sorting of tag values when multiple objects' tags are combined in `utilCombinedTags` ([#11932], thanks [@JaiswalShivang])
 * Don't change the value of `railway:turnout_side` when reversing a railway track ([#11645])
@@ -80,6 +81,7 @@ _Breaking developer changes, which may affect downstream projects or sites that 
 [#10904]: https://github.com/openstreetmap/iD/pull/10904
 [#10958]: https://github.com/openstreetmap/iD/issues/10958
 [#11052]: https://github.com/openstreetmap/iD/pull/11052
+[#11329]: https://github.com/openstreetmap/iD/issues/11329
 [#11645]: https://github.com/openstreetmap/iD/pull/11645
 [#11661]: https://github.com/openstreetmap/iD/pull/11661
 [#11778]: https://github.com/openstreetmap/iD/pull/11778